+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - EM-REML algorithm generalized from single random intercept to arbitrary unstructured covariance models
+- Replaced the Py-BOBYQA dependency and default optimizer with SciPy COBYQA
+- Reduced unused and redundant Python and Rust dependencies
 
 ## [1.1.0] - 2026-01-27